+++ src/app.tsx
@@ -25,7 +25,7 @@ const SearchReplaceForBlockEditor = () => {
   const [searchInput, setSearchInput] = useState('');
   const [replaceInput, setReplaceInput] = useState('');
   const [caseSensitive, setCaseSensitive] = useState(false);
-  const [matchCount, setMatchCount] = useState(0);
+  const [context, setContext] = useState(false);
 
   /**
    * Open Modal.
@@ -37,7 +37,7 @@ const SearchReplaceForBlockEditor = () => {
   const openModal = (): void => {
     setIsModalVisible(true);
     setReplacements(0);
-    setMatchCount(0);
+    searchInput ? replace(false) : '';
   }
 
   /**
@@ -50,7 +50,6 @@ const SearchReplaceForBlockEditor = () => {
   const closeModal = (): void => {
     setIsModalVisible(false);
     setReplacements(0);
-    setMatchCount(0);
   }
 
   /**
@@ -83,8 +82,8 @@ const SearchReplaceForBlockEditor = () => {
    * @returns {void}
    */
   useEffect(() => {
-    searchFn(searchInput);
-  }, [caseSensitive]);
+    replace(false);
+  }, [searchInput, caseSensitive]);
 
   /**
    * Handle case sensitive toggle feature
@@ -107,7 +106,8 @@ const SearchReplaceForBlockEditor = () => {
    *
    * @returns {void}
    */
-  const replace = (): void => {
+  const replace = (context = false): void => {
+    setContext(context);
     setReplacements(0);
 
     if (!searchInput) {
@@ -120,33 +120,7 @@ const SearchReplaceForBlockEditor = () => {
     );
 
     select('core/block-editor').getBlocks().forEach((element) => {
-      recursivelyReplace(element, pattern, replaceInput);
-    });
-  };
-
-  /**
-   * Handle the implementation for when the user
-   * search for a string.
-   *
-   * @since 1.3.0
-   *
-   * @returns {void}
-   */
-  const searchFn = (value): void => {
-    setMatchCount(0);
-    setSearchInput(value);
-
-    if (!value) {
-      return;
-    }
-
-    const pattern = new RegExp(
-      `(?<!<[^>]*)${value}(?<![^>]*<)`,
-      isCaseSensitive() || caseSensitive ? 'g' : 'gi'
-    );
-
-    select('core/block-editor').getBlocks().forEach((element) => {
-      recursivelySearch(element, pattern);
+      recursivelyReplace(element, pattern, replaceInput, context);
     });
   };
 
@@ -164,9 +138,9 @@ const SearchReplaceForBlockEditor = () => {
    *
    * @returns {void}
    */
-  const recursivelyReplace = (element, pattern, text) => {
+  const recursivelyReplace = (element, pattern, text, context) => {
     if (getAllowedBlocks().indexOf(element.name) !== -1) {
-      const args = { element, pattern, text };
+      const args = { element, pattern, text, context };
 
       switch (element.name) {
         case 'core/quote':
@@ -186,46 +160,7 @@ const SearchReplaceForBlockEditor = () => {
 
     if (element.innerBlocks.length) {
       element.innerBlocks.forEach((innerElement) => {
-        recursivelyReplace(innerElement, pattern, text);
-      });
-    }
-  }
-
-  /**
-   * Recursively traverse and search the text in the
-   * Block Editor with the user's text.
-   *
-   * @since 1.3.0
-   * @since 1.0.1 Handle edge-cases for quote, pullquote & details block.
-   *
-   * @param {Object} element Gutenberg editor block.
-   * @param {string} pattern Search pattern.
-   *
-   * @returns {void}
-   */
-  const recursivelySearch = (element, pattern) => {
-    if (getAllowedBlocks().indexOf(element.name) !== -1) {
-      const args = { element, pattern};
-      switch (element.name) {
-        case 'core/quote':
-        case 'core/pullquote':
-          searchBlockAttribute(args, 'citation');
-          break;
-        case 'core/details':
-          searchBlockAttribute(args, 'summary');
-          break;
-        
-        case 'core/table':
-          searchTableContent(args);
-          break;
-        default:
-          searchBlockAttribute(args, 'content');
-          break;
-      }
-    }
-    if (element.innerBlocks.length) {
-      element.innerBlocks.forEach((innerElement) => {
-        recursivelySearch(innerElement, pattern);
+        recursivelyReplace(innerElement, pattern, text, context);
       });
     }
   }
@@ -262,116 +197,17 @@ const SearchReplaceForBlockEditor = () => {
     const property = {};
     property[attribute] = newString;
 
-    (dispatch('core/block-editor') as any).updateBlockAttributes(clientId, property);
+    if(args.context){
+      (dispatch('core/block-editor') as any).updateBlockAttributes(clientId, property);
+    }
 
     // Handle edge-case ('value') with Pullquotes.
     if (attributes.value) {
-      (dispatch('core/block-editor') as any)
-        .updateBlockAttributes(clientId, { value: newString });
-      setReplacements((items) => items + 1);
-    }
-  }
-
-  /**
-   * Do the actual job of searching the string
-   * in the block editor.
-   *
-   * @since 1.3.0
-   *
-   * @param {Object} args      Args object containing element, pattern and text.
-   * @param {string} attribute The attribute to be mutated e.g. content.
-   *
-   * @returns {void}
-   */
-  const searchBlockAttribute = (args, attribute) => {
-    const { attributes } = args.element;
-    if (undefined === attributes || undefined === attributes[attribute]) {
-      return;
-    }
-    let oldString: string = attributes[attribute].text || attributes[attribute];
-    let matches = (oldString.match(args.pattern) || []).length;
-    if(matches){
-      setMatchCount(prevCount => prevCount + 1);
-    }
-  }
-
-  /**
-   * Do the actual job of searching the string in table block
-   * in the block editor.
-   *
-   * @since 1.3.0
-   *
-   * @param {Object} args Args object containing element, pattern and text.
-   *
-   * @returns {void}
-   */
-  const searchTableContent = (args) => {
-    const { attributes } = args.element;
-
-    // Handle Table Caption Replacement
-    if (attributes && attributes.caption) {
-      let oldCaptionString = attributes.caption.text || attributes.caption;
-      let matches = (oldCaptionString.match(args.pattern) || []).length;
-      if(matches){
-        setMatchCount(prevCount => prevCount + 1);
+      if(args.context){
+        (dispatch('core/block-editor') as any)
+        .updateBlockAttributes(clientId, { value: newString });  
       }
-    }
-
-    // Replace body cells content
-    if (attributes.body) {
-      const updatedBody = attributes.body.map(row => {
-        if (row.cells) {
-          row.cells = row.cells.map(cell => {
-            if (cell.content) {
-              let oldCellContent = cell.content.text || cell.content;
-              let matches = (oldCellContent.match(args.pattern) || []).length;
-              if(matches){
-                setMatchCount(prevCount => prevCount + 1);
-              }
-            }
-            return cell;
-          });
-        }
-        return row;
-      });
-    }
-
-    // Replace head cells content
-    if (attributes.head) {
-      const updatedHead = attributes.head.map(row => {
-        if (row.cells) {
-          row.cells = row.cells.map(cell => {
-            if (cell.content) {
-              let oldCellContent = cell.content.text || cell.content;
-              let matches = (oldCellContent.match(args.pattern) || []).length;
-              if(matches){
-                setMatchCount(prevCount => prevCount + 1);
-              }
-            }
-            return cell;
-          });
-        }
-        return row;
-      });
-    }
-
-    // Replace foot cells content
-    if (attributes.foot) {
-      const updatedFoot = attributes.foot.map(row => {
-        if (row.cells) {
-          row.cells = row.cells.map(cell => {
-            if (cell.content) {
-              let oldCellContent = cell.content.text || cell.content;
-              let matches = (oldCellContent.match(args.pattern) || []).length;
-              if(matches){
-                setMatchCount(prevCount => prevCount + 1);
-              }
-            }
-            return cell;
-          });
-        }
-        return row;
-      });
+      setReplacements((items) => items + 1);
     }
   }
 
@@ -421,7 +257,7 @@ const SearchReplaceForBlockEditor = () => {
                 type="text"
                 label={__('Search')}
                 value={searchInput}
-                onChange={(value) => searchFn(value)}
+                onChange={(value) => setSearchInput(value)}
                 placeholder="Lorem ipsum..."
                 __nextHasNoMarginBottom
               />
@@ -444,7 +280,7 @@ const SearchReplaceForBlockEditor = () => {
             </div>
 
             {
-              replacements ? (
+              context ? (
                 <div id="search-replace-modal__notification">
                   <p>
                     <strong>{replacements}</strong> {__('item(s) replaced successfully', 'search-replace-for-block-editor')}.
@@ -454,19 +290,19 @@ const SearchReplaceForBlockEditor = () => {
             }
 
             {
-              matchCount && !replacements ? (
+              !context && searchInput ? (
                 <div id="search-replace-modal__notification">
-                  <p>
-                    <strong>{matchCount}</strong> {__('item(s) found', 'search-replace-for-block-editor')}.
-                  </p>
-                </div>
+                <p>
+                  <strong>{replacements}</strong> {__('item(s) found', 'search-replace-for-block-editor')}.
+                </p>
+              </div>
               ) : ''
             }
 
             <div id="search-replace-modal__button-group">
               <Button
                 variant="primary"
-                onClick={replace}
+                onClick={() => replace(true)}
               >
                 {__('Replace')}
               </Button>